Exception handling - update

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -5,13 +5,14 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
 import com.educandoweb.course.services.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class UserService {
@@ -34,19 +35,28 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
+
+	    if (!userRepository.existsById(id)) {
+	        throw new ResourceNotFoundException(id);
+	    }
+
 	    try {
 	        userRepository.deleteById(id);
-	    } catch (EmptyResultDataAccessException e) {
-	        throw new ResourceNotFoundException(id);
-	    }catch(DataIntegrityViolationException e) {
-	    	throw new DatabaseException(e.getMessage());
+
+	    } catch (DataIntegrityViolationException e) {
+	        throw new DatabaseException(e.getMessage());
 	    }
 	}
 	
 	public User update(Long id,User obj) {
-		User entity = userRepository.getReferenceById(id);
-		updateData(entity, obj);
-		return userRepository.save(entity);
+		try {
+			User entity = userRepository.getReferenceById(id);
+			updateData(entity, obj);
+			return userRepository.save(entity);
+		}catch(EntityNotFoundException e) {
+			e.printStackTrace();
+			throw new ResourceNotFoundException(id);
+		}
 	}
 
 	private void updateData(User entity, User obj) {
